refac(prover): collect impl.rs to types
