+++ prover/bin/prover_client.rs
@@ -2,8 +2,7 @@ use clap::Parser;
 use jsonrpsee::http_client::{HttpClient, HttpClientBuilder};
 use jsonrpsee_core::client::ClientT;
 use jsonrpsee_core::rpc_params;
-use kroma_prover::get_proof_impl::ProofResult;
-use kroma_prover::spec_impl::SpecResult;
+use kroma_prover::types::{ProofResult, SpecResult};
 use kroma_utils::utils::b256_from_str;
 use kroma_witnessgen::get_witness_impl::WitnessResult;
 use kroma_witnessgen::request_witness_impl::RequestResult;
