+++ prover/src/lib.rs
@@ -1,6 +1,5 @@
 pub mod errors;
-pub mod get_proof_impl;
 pub mod interface;
 pub mod proof_db;
-pub mod request_prove_impl;
-pub mod spec_impl;
+pub mod types;
+pub mod utils;
