--- prover/src/request_prove_impl.rs
@@ -1,9 +1,0 @@
-use serde::{Deserialize, Serialize};
-
-#[derive(Clone, Debug, Serialize, Deserialize, PartialEq)]
-pub enum RequestResult {
-    None,
-    Processing,
-    Completed,
-    Failed(String),
-}
