+++ prover/src/types.rs
@@ -1,7 +1,43 @@
 use alloy_primitives::B256;
+use kroma_utils::{deps_version::SP1_SDK_VERSION, utils::u32_to_u8};
+use once_cell::sync::Lazy;
 use serde::{Deserialize, Serialize};
+use sp1_sdk::{HashableKey, ProverClient};
 
-use crate::{request_prove_impl::RequestResult, spec_impl::VKEY_HASH};
+// TODO: integrate elf and vkey_hash
+pub const SINGLE_BLOCK_ELF: &[u8] = include_bytes!("../../program/elf/fault-proof-elf");
+pub static VKEY_HASH: Lazy<B256> = Lazy::new(|| {
+    let prover = ProverClient::new();
+    let (_, vkey) = prover.setup(SINGLE_BLOCK_ELF);
+    B256::from(u32_to_u8(vkey.vk.hash_u32()))
+});
+
+#[derive(Clone, Debug, Serialize, Deserialize)]
+pub struct SpecResult {
+    pub version: String,
+    pub sp1_version: String,
+    pub vkey_hash: B256,
+}
+
+impl SpecResult {
+    pub fn new(version: String, sp1_version: String, vkey_hash: B256) -> Self {
+        Self { version, sp1_version, vkey_hash }
+    }
+}
+
+impl Default for SpecResult {
+    fn default() -> Self {
+        SpecResult::new("0.1.0".to_string(), SP1_SDK_VERSION.to_string(), *VKEY_HASH)
+    }
+}
+
+#[derive(Clone, Debug, Serialize, Deserialize, PartialEq)]
+pub enum RequestResult {
+    None,
+    Processing,
+    Completed,
+    Failed(String),
+}
 
 #[derive(Clone, Debug, Serialize, Deserialize, PartialEq)]
 pub struct ProofResult {
